Refactor code and remove unused classes and imports
